control durations more carefully
Instead of playing every note for the default one second, figure out how
long it should be playing and play it that long.

Additionally, for the harmony, don't re-attack chords every time the
melody notes attack; instead, wait for a chord change (or the start of a
new phrase) to re-attack.